+++ db/migrate/20230217155419_add_host_meeting_to_involuntary_conversion_projects.rb
@@ -1,0 +1,5 @@
+class AddHostMeetingToInvoluntaryConversionProjects < ActiveRecord::Migration[7.0]
+  def change
+    add_column :conversion_involuntary_task_lists, :stakeholder_kick_off_meeting, :boolean
+  end
+end