+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_02_15_164530) do
+ActiveRecord::Schema[7.0].define(version: 2023_02_17_160251) do
   create_table "contacts", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.uuid "project_id"
     t.string "name", null: false
@@ -127,6 +127,8 @@ ActiveRecord::Schema[7.0].define(version: 2023_02_15_164530) do
     t.boolean "deed_of_variation_signed_secretary_state"
     t.boolean "direction_to_transfer_not_applicable"
     t.date "stakeholder_kick_off_confirmed_conversion_date"
+    t.boolean "stakeholder_kick_off_meeting"
+    t.boolean "stakeholder_kick_off_check_provisional_conversion_date"
   end
 
   create_table "conversion_voluntary_task_lists", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|